Refactored Rooms and Room.
